add fe_low_mem.c to wolfssl vs project and fix a couple warnings

+++ src/wolfio.c
@@ -1030,7 +1030,7 @@ int wolfIO_TcpBind(SOCKET_T* sockfd, word16 port)
 #ifdef HAVE_SOCKADDR
 int wolfIO_TcpAccept(SOCKET_T sockfd, SOCKADDR* peer_addr, XSOCKLENT* peer_len)
 {
-    return accept(sockfd, peer_addr, peer_len);
+    return (int)accept(sockfd, peer_addr, peer_len);
 }
 #endif /* HAVE_SOCKADDR */
